+++ packages/firebase-utils/src/firestore/index.ts
@@ -109,7 +109,8 @@ export {
   updateScene,
   deleteScene,
   getEventScenes,
-  getOrganizationTemplateScenes,
+  getBrandScenes,
+  getOrganizationScenes,
   duplicateScene,
 } from './scenes';